// React/Site/src/API.ts
export type User = {
  __typename: "User",
  id: string,
  fullName: string,
  firebaseUid: string,
  email: string,
  displayName: string,
  imageUri?: string | null,
  defaultMidiDevice?: string | null,
  purchasedTokens: number,
  remainingTokens: number,
  totalStreams: number,
  totalSeconds: number,
  lastStream?: string | null,
  createdInstanceId?: string | null,
  hostRoomUser?: ModelHostRoomUserConnection | null,
  createdAt: string,
  updatedAt: string,
};

export type ModelHostRoomUserConnection = {
  __typename: "ModelHostRoomUserConnection",
  items:  Array<HostRoomUser | null >,
  nextToken?: string | null,
};

export type HostRoomUser = {
  __typename: "HostRoomUser",
  id: string,
  userId: string,
  hostRoomId: string,
  lastSeen?: string | null,
  archived: boolean,
  pinned: boolean,
  user?: User | null,
  hostRoom?: HostRoom | null,
  createdAt: string,
  updatedAt: string,
  userHostRoomUserId?: string | null,
  hostRoomHostRoomUsersId?: string | null,
  hostRoomUserUserId?: string | null,
};

export type HostRoom = {
  __typename: "HostRoom",
  id: string,
  name: string,
  desc?: string | null,
  imageUri?: string | null,
  createdUserId: string,
  lastMessageId?: string | null,
  lastMessage?: HostRoomChatMessage | null,
  chatMessages?: ModelHostRoomChatMessageConnection | null,
  hostRoomUsers?: ModelHostRoomUserConnection | null,
  createdAt: string,
  updatedAt: string,
  hostRoomLastMessageId?: string | null,
};
// ...
export type OnUpdateUserSubscription = {
  onUpdateUser?:  {
    __typename: "User",
    id: string,
    fullName: string,
    firebaseUid: string,
    email: string,
    displayName: string,
    imageUri?: string | null,
    defaultMidiDevice?: string | null,
    purchasedTokens: number,
    remainingTokens: number,
    totalStreams: number,
    totalSeconds: number,
    lastStream?: string | null,
    createdInstanceId?: string | null,
    hostRoomUser?:  {
      __typename: "ModelHostRoomUserConnection",
      items:  Array< {
        __typename: "HostRoomUser",
        id: string,
        userId: string,
        hostRoomId: string,
        lastSeen?: string | null,
        archived: boolean,
        pinned: boolean,
        createdAt: string,
        updatedAt: string,
        userHostRoomUserId?: string | null,
        hostRoomHostRoomUsersId?: string | null,
        hostRoomUserUserId?: string | null,
      } | null >,
      nextToken?: string | null,
    } | null,
    createdAt: string,
    updatedAt: string,
  } | null,
};